% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_flexible_image_patch.R
\name{cimg_to_spatrast}
\alias{cimg_to_spatrast}
\title{Convert a cimg object to a SpatRaster, preserving RGB metadata}
\usage{
cimg_to_spatrast(img, normalize = FALSE)
}
\arguments{
\item{img}{A `cimg` object (any number of channels).}

\item{normalize}{Logical.  If `TRUE`, rescale values to [0, 1].}
}
\value{
A `SpatRaster` with RGB metadata set when the image has 3 or 4
  channels.
}
\description{
Internal helper that replaces the bare `terra::rast(as.array(img))` pattern
used throughout the package.  The key addition is `terra::RGB(r) <- ...`,
which registers the channel indices so that `terra::plotRGB()` works without
the user having to know about this terra requirement.
}
\keyword{internal}
