% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_binary_image.R
\name{clean_root_image}
\alias{clean_root_image}
\title{Process root images by filling holes and optionally smoothing edges}
\usage{
clean_root_image(
  img,
  remove_small_particles = TRUE,
  min_size = 100,
  connectivity = 8,
  fill_holes = TRUE,
  min_hole_size = 1,
  max_hole_size = Inf,
  smooth_edges = FALSE,
  kernel_shape = "disk",
  kernel_size = 3,
  iterations = 1
)
}
\arguments{
\item{img}{A binary image of root systems (imager cimg object)}

\item{remove_small_particles}{Boolean indicating whether only big objects are retained}

\item{min_size}{Minimum particle size (in pixel). Removing all patches below the minimal size}

\item{connectivity}{Determines which connection type is used to assess whether pixel belong to the same patch: 4 or 8}

\item{fill_holes}{Boolean indicating whether to fill holes}

\item{min_hole_size}{Minimum size of holes to fill (in pixels)}

\item{max_hole_size}{Maximum size of holes to fill (in pixels, Inf means no limit)}

\item{smooth_edges}{Boolean indicating whether to smooth edges}

\item{kernel_shape}{Shape of the kernel for edge smoothing: "square", "diamond", or "disk"}

\item{kernel_size}{Size of the kernel for edge smoothing (odd integer)}

\item{iterations}{Number of iterations for the smoothing process}
}
\value{
A processed binary image
}
\description{
Process root images by filling holes and optionally smoothing edges
}
\examples{
data("seg_Oulanka2023_Session01_T067")
img <-  seg_Oulanka2023_Session01_T067
# Fill holes only
result1 <- clean_root_image(img, fill_holes = TRUE, smooth_edges = FALSE)
# Fill holes and smooth edges
result2 <- clean_root_image(img, fill_holes = TRUE, smooth_edges = TRUE)
plot(result2)
}
\seealso{
\code{\link{smooth_root_edges}}, \code{\link{fill_root_holes}}, \code{\link{remove_small_foreground_objects}}
}
