% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_image.R
\name{.plot_size_hist}
\alias{.plot_size_hist}
\title{Histogram of component sizes on a log10 axis, with mean/median marked}
\usage{
.plot_size_hist(sizes, main, breaks)
}
\description{
Histogram of component sizes on a log10 axis, with mean/median marked
}
\keyword{internal}
