% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calibration3.R
\name{estimate_rotation_center}
\alias{estimate_rotation_center}
\title{Estimates rotation from tape coverage}
\usage{
estimate_rotation_center(
  img,
  tape_brightness = 0.66,
  extra_rows = 100,
  search_area = 0.45,
  tape_quantile = 0.98,
  nclasses = 3,
  select_layer = NULL
)
}
\arguments{
\item{img}{Input image as raster, file name, or array}

\item{tape_brightness}{Brightness threshold for tape detection (0-1)}

\item{extra_rows}{Additional rows to add for analysis}

\item{search_area}{Proportion of image to analyze (0-1)}

\item{tape_quantile}{Quantile used to align brightness with tape (0-1)}

\item{nclasses}{Number of classes for pixel clustering}

\item{select_layer}{Integer. Specifies which layer to use if the input is a multi-band image. Default is `NULL`.}
}
\value{
numeric Position of the center of extruding tape
}
\description{
This function analyzes image data to determine rotation based on tape coverage,
assuming more tape is present on the upper side of the tube.
}
\examples{
img = seg_Oulanka2023_Session01_T067
r0 = estimate_rotation_center(img)
}
