% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calibration3.R
\name{estimate_soil_surface}
\alias{estimate_soil_surface}
\title{Estimate soil surface position using tape markers}
\usage{
estimate_soil_surface(
  img,
  search_area = 0.45,
  tape_thresh = 0.33,
  dpi = 150,
  nclasses = 3,
  inverse = FALSE,
  tape_overlap = 0.5,
  tape_brightness = 0.6,
  extra_rows = 100,
  tape_quantile = 0.98,
  select_layer = NULL
)
}
\arguments{
\item{img}{Input image (raster, filename, or array)}

\item{search_area}{Proportion of image to analyze}

\item{tape_thresh}{Minimum tape coverage ratio}

\item{dpi}{Image resolution}

\item{nclasses}{Number of clustering classes}

\item{inverse}{Invert detection for dark markers}

\item{tape_overlap}{Safety margin for tape (cm)}

\item{tape_brightness}{Brightness threshold for tape}

\item{extra_rows}{Additional analysis rows}

\item{tape_quantile}{Brightness alignment quantile}

\item{select_layer}{Integer. Specifies which layer to use if the input is a multi-band image. Default is `NULL`.}
}
\value{
data.frame with soil surface and tape end positions
}
\description{
Detects the soil surface by analyzing tape coverage patterns in the image.
}
\examples{
img = rgb_Oulanka2023_Session03_T067
Soil0Estimates = estimate_soil_surface(img)
}
