% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_binary_image.R
\name{fill_holes}
\alias{fill_holes}
\title{Fill holes in binary images}
\usage{
fill_holes(img, max_size = NULL)
}
\arguments{
\item{img}{A `cimg` object representing a binary image (values 0 and 1).}

\item{max_size}{Optional maximum size (in pixels) of holes to fill. If `NULL`, all holes are filled.}
}
\value{
A `cimg` object with holes filled (as 1s).
}
\description{
Identifies internal black regions ("holes") in a binary image and fills them by setting their pixel values to 1. Holes are defined as black areas (value = 0) completely surrounded by white (value = 1), i.e., not connected to the image border.
}
\keyword{internal}
