% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_binary_image.R
\name{fill_root_holes}
\alias{fill_root_holes}
\title{Fill internal holes in root binary images while preserving edges}
\usage{
fill_root_holes(img, min_hole_size = 1, max_hole_size = Inf)
}
\arguments{
\item{img}{A binary image of root systems (imager cimg object)}

\item{min_hole_size}{Minimum size of holes to fill (in pixels)}

\item{max_hole_size}{Maximum size of holes to fill (in pixels, Inf means no limit)}
}
\value{
A processed binary image with internal holes filled
}
\description{
Fill internal holes in root binary images while preserving edges
}
\examples{
\dontrun{
data("seg_Oulanka2023_Session01_T067")
img <- seg_Oulanka2023_Session01_T067
filled <- fill_root_holes(img, min_hole_size = 1, max_hole_size = 100)
plot(filled)
}
}
\keyword{internal}
