% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rhythmicity.R
\name{fit_sine_curve}
\alias{fit_sine_curve}
\title{Fit a sine curve using non-linear least squares}
\usage{
fit_sine_curve(
  x,
  y,
  parStart = list(amp = 3, phase = 0, offset = 0, period = 24)
)
}
\arguments{
\item{x}{Numeric vector of time values (e.g., hours or days).}

\item{y}{Numeric vector of response values.}

\item{parStart}{Named list with starting values for parameters: `amp`, `phase`, `offset`, and `period`.}
}
\value{
A named list with:
\describe{
  \item{amp}{Amplitude (peak distance from baseline).}
  \item{phase}{Phase shift (horizontal time shift/ scanner rotation) in same units as `tt`.}
  \item{offset}{Vertical shift (baseline).}
  \item{peak}{Time position of the peak / scanner rotation position, relative to phase and period.}
  \item{period}{Distance between peaks}
  \item{tss}{Total sum of squares.}
  \item{rss}{Residual sum of squares.}
  \item{R2}{Coefficient of determination (pseudo R²).}
  \item{residual_se}{Residual standard error.}
  \item{df_residual}{Degrees of freedom of the residual.}
  \item{predicted}{Fitted values.}
  \item{residuals}{Model residuals.}
}
}
\description{
This function fits a sine curve of the form  
\eqn{y = \text{amp} \cdot \sin\left(\frac{2\pi}{\text{period}} \cdot (x + \text{phase})\right) + \text{offset}}  
to time series data using Levenberg-Marquardt optimization from `minpack.lm::nls.lm`.
}
