% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LilHelpers3.R
\name{image_threshold}
\alias{image_threshold}
\title{Threshold or deblur an image to binarize features}
\usage{
image_threshold(
  img,
  threshold = 0.4,
  method = "global",
  window_size = 15,
  select_layer = 2,
  mask_layer = 1,
  binary_01 = FALSE,
  deblur = FALSE
)
}
\arguments{
\item{img}{SpatRaster object}

\item{threshold}{Numeric value (0–1). For global thresholding, it's a fraction of the global max.
For adaptive thresholding, it's a fraction of local mean. For deblurring, it's the fraction of max used to recover structure.}

\item{method}{"global" or "adaptive". Ignored if `deblur = TRUE`.}

\item{window_size}{Integer (odd), only used for adaptive thresholding.}

\item{select_layer}{Integer or NULL. Which layer to use for thresholding or deblurring.
If NULL and multilayer, the mean of all layers is used.}

\item{mask_layer}{Integer or NULL. If set, used to preserve masked regions or enhance structure.}

\item{binary_01}{Logical. If TRUE, binarized output uses 0/1. If FALSE, it retains max value of input.}

\item{deblur}{Logical. If TRUE, applies deblurring logic instead of standard thresholding.}
}
\value{
SpatRaster object
}
\description{
Applies global or adaptive thresholding to a raster image, with optional
masking and simple deblurring based on structural layer separation.
}
\examples{
img = terra::rast(seg_Oulanka2023_Session03_T067)
image_threshold(img, threshold = 0.3, method = "global")
image_threshold(img, threshold = 0.9, method = "adaptive", window_size = 15, binary_01 = TRUE)
image_threshold(img, threshold = 0.4, select_layer = 2, mask_layer = 1, deblur = TRUE)
}
