% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexible_inputs.R
\name{load_flexible_image}
\alias{load_flexible_image}
\title{Load an image flexibly from file or convert from memory}
\usage{
load_flexible_image(
  input,
  output_format = "cimg",
  normalize = TRUE,
  select.layer = NULL,
  binarize = FALSE
)
}
\arguments{
\item{input}{File path or image object}

\item{output_format}{Character, one out of "cimg", "spatrast", "matrix", "array", "brick", "raster", "spatrast", "magick-image". Other spellings are accepted.}

\item{normalize}{Logical, whether to normalize values to 0-1 range if they're in 0-255}

\item{select.layer}{Numeric, which layer to select if input has multiple layers}

\item{binarize}{Logical, whether the output is strictly 0 and 1. Overwrites normalize}
}
\description{
Load an image flexibly from file or convert from memory
}
