% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareDistributions3.R
\name{MRD}
\alias{MRD}
\title{Calculate Mean Rooting Depth}
\usage{
MRD(w, roots)
}
\arguments{
\item{w}{Numeric vector of weights (typically depths)}

\item{roots}{Numeric vector of root coverage values}
}
\value{
Numeric MRD value
}
\description{
Calculate Mean Rooting Depth
}
\examples{
w <- seq(5, 25, 5)
roots <- c(0, 10, 7, 3, 1)
mean_rooting_depth <- MRD(w, roots)
}
