% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_binary_image.R
\name{remove_small_foreground_objects}
\alias{remove_small_foreground_objects}
\title{Remove small foreground objects (e.g., dirt) from root binary images}
\usage{
remove_small_foreground_objects(img, min_size = 100, connectivity = 8)
}
\arguments{
\item{img}{A binary image of root systems (imager cimg object)}

\item{min_size}{Minimum size in pixels for foreground objects to keep}

\item{connectivity}{Type of connectivity for connected components: 4 or 8 (default: 8)}
}
\value{
A processed binary image with small foreground objects removed (0=background, 1=foreground)
}
\description{
Remove small foreground objects (e.g., dirt) from root binary images
}
\examples{
\dontrun{
data("seg_Oulanka2023_Session01_T067")
img <-  seg_Oulanka2023_Session01_T067
cleaned <- remove_small_foreground_objects(img, min_size = 100)
plot(cleaned)
}
}
\keyword{internal}
