% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rgb_Oulanka2023_Session03_T067}
\alias{rgb_Oulanka2023_Session03_T067}
\title{Original Minirhizotron Root Scan - Session 3, Tube 67}
\format{
A RasterBrick object with dimensions:
  \itemize{
    \item 4900 columns (width)
    \item 1161 rows (height)
    \item 3 layers (RGB channels)
  }
}
\source{
Original minirhizotron images acquired by Johannes Cunow and Gesche Blume-Werry
at Oulanka Research Station, Finland, in October 2023.
}
\usage{
data(rgb_Oulanka2023_Session03_T067)
}
\description{
Original RGB root scan image from a sedge fen in northern Finland (October 2023).
This image represents a composite of multiple stitched scans.
}
\details{
The image represents a complete tube scan where:
\itemize{
  \item Columns correspond to tube length
  \item Rows correspond to tube rotation
  \item RGB channels represent true color information
}
}
\examples{
\dontrun{
  data(rgb_Oulanka2023_Session03_T067)
  rgb_Oulanka2023_Session03_T067 = terra::rast(rgb_Oulanka2023_Session03_T067)
  terra::plotRGB(rgb_Oulanka2023_Session03_T067, maxcell = Inf)
}
}
\author{
Johannes Cunow \email{johannes.cunow@gmail.com}
}
\keyword{datasets}
