% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rhythmicity.R
\name{rhytmicity_test}
\alias{rhytmicity_test}
\title{Finite sample/ large sample Likelihood ratio test for circadian pattern detection}
\usage{
rhytmicity_test(
  tt,
  yy,
  parStart = list(amp = 3, phase = 0, offset = 0, period = 12),
  method = "FRL",
  FN = TRUE
)
}
\arguments{
\item{tt}{Time vector / scanner rotation position}

\item{yy}{Value}

\item{parStart}{Named list with starting values for parameters: `amp`, `phase`, `offset`, and `period`.}

\item{method}{Type of test to use, `FLR` (finite sample liklihood ratio test) for a F-like Test, or `LR` (large samples) for liklihood ratio test, or `F` for a standard F-test. Default is finite sample liklhood ratio test.}
}
\value{
A list of amp, phase, offset, sigma02, sigmaA2, l0, l1, df, stat, and pvalue.
model: \eqn{y= A * sin(2*pi*x+B)+C}
\item{y}{a 1*n vector of data y}
\item{A}{estimated A^hat from fit_sine_curve}
\item{B}{estimated B^hat from fit_sine_curve}
\item{C}{estimated C^hat from fit_sine_curve}
\item{sigma0}{sigma0^hat under H0}
\item{sigmaA}{sigmaA^hat under H1}
\item{n}{length of data y}
\item{df0}{df under H0}
\item{df1}{df under H1}

 \item{amp}{Amplitude based on formula 1}
 \item{phase}{Phase based on formula 1, phase is restricted within (0, period)}
 \item{peakTime}{Phase based on formula 1, peakTime is restricted within (0, period). phase + peakTime = period/4}
 \item{offset}{Basal level(vertical shift) based on formula 1 or on formula 2}
 \item{sigma02}{Variance estimate under the null (intercept only)}
 \item{sigmaA2}{Variance estimate under the alternative (since curve fitting)}
 \item{l0}{Log likelihood under the null (intercept only)}
 \item{l1}{Log likelihood under the alternative (since curve fitting)}
 \item{df}{Degree of freedom for the LR test}
 \item{stat}{LR statistics}
 \item{pvalue}{P-value from the LR test}
 \item{R2}{Pseudo R2 defined as (tss - rss)/tss}
}
\description{
Test the significance of circadian curve fitting using finite sample likelihood ratio test
}
\details{
Formula 1: \eqn{yy = amp \times sin(2\pi/period \times (phase + tt)) + offset}
Formula 2: \eqn{yy = A \times sin(2\pi/period \times tt) + B * cos(2*pi/period * tt) + offset}

sourced from: https://rdrr.io/github/diffCircadian/diffCircadian/ under GPL3 license
author: Zhiguang Huo, Haocheng Ding
Feb. 24, 2023, 9:07 a.m.; Version 0.0.0
journal article: https://doi.org/10.1093/bib/bbab224
tutorial (for circadian data): http://htmlpreview.github.io/?https://github.com/diffCircadian/diffCircadian/blob/master/vignettes/diffCircadian_tutorial.html
}
\examples{
set.seed(32608)
n <- 50
Period <- 24
tt <- runif(n,0,Period)
Amp <- 2
Phase <- 6
Offset <- 3
yy <- Amp * sin(2*pi/Period * (tt + Phase)) + Offset + rnorm(n,0,1)
rhytmicity_test(tt, yy,  method = "FRL", 
                parStart = list(amp = 3, phase = 0, offset = 0, period = 20) )
}
\author{
Caleb (Zhiguang Huo)
}
\keyword{internal}
