% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zonestats3.R
\name{root_thickness}
\alias{root_thickness}
\title{Approximate average Root Thickness}
\usage{
root_thickness(kimuralength, rootpx, dpi = 300)
}
\arguments{
\item{kimuralength}{Total root length in cm (e.g. from [root_length()]).}

\item{rootpx}{Total number of root pixels in the image section.}

\item{dpi}{Image resolution in dots per inch. Default is 300.}
}
\value{
A numeric value in cm representing approximate average root diameter.
}
\description{
Estimates mean root diameter from total root length and root pixel count.
}
\examples{
root.thicc <- root_thickness(kimuralength = 300, rootpx = 9500, dpi = 300)
}
