% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seg_Oulanka2023_Session01_T067}
\alias{seg_Oulanka2023_Session01_T067}
\title{Segmented Minirhizotron Root Scan - Session 1, Tube 67}
\format{
A RasterBrick object with dimensions:
  \itemize{
    \item 4900 columns (width)
    \item 1144 rows (height)
    \item 3 layers (channels)
  }
}
\source{
Original minirhizotron images acquired by Johannes Cunow and Gesche Blume-Werry
at Oulanka Research Station, Finland, in June 2023.
}
\usage{
data(seg_Oulanka2023_Session01_T067)
}
\description{
Segmented root scan image from a sedge fen in northern Finland (June 2023).
The image was processed using RootDetector for root segmentation.
}
\details{
Binary mask representation where:
\itemize{
  \item Roots = 1
  \item Background = 0
  \item Layer 1 includes foreign objects (e.g., tape) marked as 1
}
Spatial dimensions correspond to physical tube measurements:
columns = tube length, rows = tube rotation.
}
\examples{
\dontrun{
  data(seg_Oulanka2023_Session01_T067)
  seg_Oulanka2023_Session01_T067 = terra::rast(seg_Oulanka2023_Session01_T067)
  terra::plot(seg_Oulanka2023_Session01_T067, maxcell = Inf)
}
}
\author{
Johannes Cunow \email{johannes.cunow@gmail.com}
}
\keyword{datasets}
