% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thinning4.R
\name{thin_image_steepest_ascend}
\alias{thin_image_steepest_ascend}
\title{Steepest Ascend Skeletonization (Internal)}
\usage{
thin_image_steepest_ascend(img, verbose = FALSE, select.layer = NULL)
}
\arguments{
\item{img}{A matrix, data frame, or `SpatRaster` object representing the binary image.}

\item{verbose}{Logical. If `TRUE`, prints diagnostic information. Default is `FALSE`.}

\item{select.layer}{Integer indicating the layer to use if `img` is a multi-layer `SpatRaster`. Default is `NULL`.}
}
\value{
A binary matrix representing the skeleton obtained from local maxima on the distance transform.
}
\description{
Skeletonize a binary image using a ridge-based steepest ascend approximation on the distance transform.
}
\keyword{internal}
