% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Turnover3.R
\name{turnover_tc}
\alias{turnover_tc}
\title{Calculate global root production and root turnover from temporal comparison}
\usage{
turnover_tc(
  im.t1,
  im.t2,
  method = "kimura",
  unit = "cm",
  dpi = 300,
  select.layer = 2
)
}
\arguments{
\item{im.t1}{SpatRaster object for timepoint 1}

\item{im.t2}{SpatRaster object for timepoint 2}

\item{method}{Analysis method: "kimura" or "rootpx"}

\item{unit}{Unit of root length measurement (only for method = "kimura"). Default: "cm"}

\item{dpi}{Image resolution (only for method = "kimura"). Default: 300}

\item{select.layer}{Integer specifying the layer to use in both timesteps if \code{img} is a multi-layer `SpatRaster`. Defaults to 2.}
}
\value{
data.frame containing:
  - standingroot_t1: Standing roots at first timepoint
  - standingroot_t2: Standing roots at second timepoint
  - production: Root production between timepoints
  - newroot%per_t1: Percentage of new roots compared to starting conditions
  - newroot%per_t2: Percentage of new roots at second timepoint
}
\description{
Calculate global root production and root turnover from temporal comparison
}
\examples{
  data(skl_Oulanka2023_Session01_T067)
  data(skl_Oulanka2023_Session03_T067)
  time1 <- terra::rast(skl_Oulanka2023_Session01_T067)
  time2 <- terra::rast(skl_Oulanka2023_Session03_T067)
  turnover.values <- turnover_tc(
    im.t1 = time1,
    im.t2 = time2,
    method = "kimura")
}
\keyword{internal}
