% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TurnoverDPC_data}
\alias{TurnoverDPC_data}
\title{Root Turnover Analysis Data}
\format{
A RasterBrick object with dimensions:
  \itemize{
    \item 2550 columns (width)
    \item 2273 rows (height)
    \item 3 layers (decay, growth, persistent)
  }
}
\source{
Images by J.Cunow
}
\usage{
data(TurnoverDPC_data)
}
\description{
Root turnover analysis from a sedge fen in northern Finland, comparing
root presence between two time points (June 2023 and October 2023) using
RootDetector root tracking feature.
}
\details{
Multi-layer representation of root dynamics where:
\itemize{
  \item Layer 1: Root decay (value = 1)
    \itemize{
      \item Roots that disappeared between time points
      \item Foreign objects (e.g., tape)
      \item Persistent roots
    }
  \item Layer 2: Root growth (value = 1)
    \itemize{
      \item New roots that appeared
      \item Persistent roots
    }
  \item Layer 3: Persistent roots (value = 1)
    \itemize{
      \item Roots present in both time points
    }
}
Background is represented as 0 in all layers.
}
\examples{
\dontrun{
  data(TurnoverDPC_data)
  img = terra::rast(TurnoverDPC_data)
  # Plot individual layers
  show_root_image(img, fun = "mean")
  # Calculate turnover statistics
  decay <- sum(terra::values(img[[1]]) == 1)
  growth <- sum(terra::values(img[[2]]) == 1)
  persistent <- sum(terra::values(img[[3]]) == 1)
}
}
\author{
Johannes Cunow \email{johannes.cunow@gmail.com}
}
\keyword{datasets}
