% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateDepthmap3.R
\name{create_depthmap}
\alias{create_depthmap}
\title{Create A Phase-Shifted, Tilt-Amplitude Sine Depth Map}
\usage{
create_depthmap(
  img,
  mask = NULL,
  sinoid = TRUE,
  tube_thicc = 7,
  tilt = 45,
  dpi = 300,
  start_soil = 0,
  center_offset = 0.5,
  progress = FALSE
)
}
\arguments{
\item{img}{Input image (accepts terra SpatRaster, matrix, array, or file path). For multi-band
images, specify band_index parameter}

\item{mask}{Raster mask indicating foreign objects (1 = mask, 0 or NA = keep)}

\item{sinoid}{Logical; if TRUE, accounts for tube curvature in depth calculation}

\item{tube_thicc}{Numeric; diameter of minirhizotron tube in cm}

\item{tilt}{Numeric; minirhizotron tube insertion angle in degrees (typically 30-45)}

\item{dpi}{Numeric; image resolution in dots per inch}

\item{start_soil}{Numeric; soil surface boundary in cm (0 = surface)}

\item{center_offset}{Numeric; rotational center offset (0 = centered, 1 = edge)}

\item{progress}{Message; indicates how mny rows have been processed}
}
\value{
terra raster object containing the depth map
}
\description{
This function generates a depth map for minirhizotron images, accounting for tube geometry
and insertion angle.
}
\examples{
data(seg_Oulanka2023_Session01_T067)
seg_Oulanka2023_Session01_T067 = terra::rast(seg_Oulanka2023_Session01_T067)
img = seg_Oulanka2023_Session01_T067
mask = seg_Oulanka2023_Session01_T067[[1]] - seg_Oulanka2023_Session01_T067[[2]]
mask[mask == 255] <- NA
map = create_depthmap(img,mask,start_soil = 0.1,
  sinoid = TRUE,
  tube_thicc = 7,
  tilt = 45,
  dpi = 300,
  center_offset = 0.1 )
}
\author{
Johannes Cunow, Robert Weigel
}
