% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_image.R
\name{remove_small_objects}
\alias{remove_small_objects}
\title{Remove small isolated white artifacts from a binary image}
\usage{
remove_small_objects(img, max_size = NULL)
}
\arguments{
\item{img}{A `cimg` binary image (values 0 and 1).}

\item{max_size}{Maximum artifact size in pixels to remove.  If `NULL`
(default), all isolated white regions are removed.}
}
\value{
A `cimg` with artifacts removed.
}
\description{
Sets to 0 all white regions (value = 1) that are not connected to the
image border.
}
\keyword{internal}
