% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calibration3.R
\name{rotation_censor}
\alias{rotation_censor}
\title{Censor image edges based on rotation}
\usage{
rotation_censor(
  img,
  center.offset = 0,
  cut.buffer = 0.02,
  fixed.rotation = TRUE,
  fixed.width = 500,
  select.layer = NULL
)
}
\arguments{
\item{img}{Input image to censor}

\item{center.offset}{Rotation shift in rows (from estimate_rotation_shift())}

\item{cut.buffer}{Proportion of image to cut when fixed_rotation=FALSE}

\item{fixed.rotation}{Use fixed output dimensions}

\item{fixed.width}{Output width when fixed_rotation=TRUE}

\item{select.layer}{Integer. Specifies which layer to use if the input is a multi-band image. Default is `NULL`.}
}
\value{
Cropped raster image
}
\description{
Crops image edges to handle non-overlapping regions between sequential scans.
}
\examples{
data(seg_Oulanka2023_Session01_T067)
img = terra::rast(seg_Oulanka2023_Session01_T067)
censored.raster = rotation_censor(img,
                         center.offset = 120,
                         cut.buffer = 0.02,
                         fixed.rotation = FALSE)
                         
censored.raster = rotation_censor(img,
                         center.offset = 220,
                         cut.buffer = 0.02,
                         fixed.width = 1000,
                         fixed.rotation = TRUE)
}
