% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thinning4.R
\name{skeletonize_image}
\alias{skeletonize_image}
\title{Skeletonization Wrapper Function}
\usage{
skeletonize_image(
  img,
  methods = c("ZhangSuen", "GuoHall", "MAT", "SteepestAscend"),
  verbose = TRUE,
  select.layer = NULL
)
}
\arguments{
\item{img}{A matrix, data frame, or `SpatRaster` object representing the binary image to be skeletonized.}

\item{methods}{A character vector specifying the skeletonization methods to apply. Valid options are \code{"ZhangSuen"}, \code{"GuoHall"}, and \code{"MAT"}. Defaults to all three methods.}

\item{verbose}{Logical. If \code{TRUE}, displays progress and diagnostic messages during processing. Defaults to \code{TRUE}.}

\item{select.layer}{Integer specifying the layer to use if \code{img} is a multi-layer `SpatRaster`. Defaults to 2.}
}
\value{
If a single method is selected, the function returns a `SpatRaster` object representing the skeletonized image. If multiple methods are selected, a named list of `SpatRaster` objects is returned.
}
\description{
This function serves as a wrapper for applying different skeletonization methods to a binary image, including the Zhang-Suen, Guo-Hall, and Medial Axis Transform (MAT) algorithms.
}
\details{
This function allows for flexible and streamlined skeletonization of binary images using one or more supported algorithms:
\itemize{
  \item \code{"ZhangSuen"}: Implements the Zhang-Suen thinning algorithm.
  \item \code{"GuoHall"}: Implements the Guo-Hall thinning algorithm.
  \item \code{"MAT"}: Computes the Medial Axis Transform to extract the skeleton.
}

The function processes the input image with the specified methods and returns the results. If multiple methods are chosen, the results are returned as a named list, with each element corresponding to a method.
}
\examples{

# Load a binary image as a SpatRaster
binary_image <- terra::rast(matrix(c(0, 1, 1, 0, 0, 1, 1, 0,0), nrow = 3))

# Apply all skeletonization methods
skeletons <- skeletonize_image(binary_image, verbose = TRUE)


}
\seealso{
\code{\link{thin_image_zhangsuen}}, \code{\link{thin_image_guohall}}, \code{\link{medial_axis_transform}}
}
