% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{skl_Oulanka2023_Session03_T067}
\alias{skl_Oulanka2023_Session03_T067}
\title{Skeletonized Root Scan - Session 3, Tube 67}
\format{
A RasterBrick object with dimensions:
  \itemize{
    \item 4900 columns (width)
    \item 1161 rows (height)
    \item 3 layers (channels)
  }
}
\source{
Images by J.Cunow
}
\usage{
data(skl_Oulanka2023_Session03_T067)
}
\description{
Skeletonized root scan image from a sedge fen in northern Finland (October 2023).
The image was processed using RootDetector for segmentation and skeletonization.
}
\details{
Binary mask representation where:
\itemize{
  \item Root skeletons = 1
  \item Background = 0
  \item Layer 1 includes foreign objects (e.g., tape) marked as 1
}
Skeletonization reduces root width to single-pixel lines while preserving
the root system topology.
}
\examples{
\dontrun{
  data(skl_Oulanka2023_Session03_T067)
  skl_Oulanka2023_Session03_T067 = terra::rast(skl_Oulanka2023_Session03_T067)
  terra::plot(skl_Oulanka2023_Session03_T067, maxcell = Inf)
}
}
\author{
Johannes Cunow \email{johannes.cunow@gmail.com}
}
\keyword{datasets}
