% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_image.R
\name{smooth_root_edges}
\alias{smooth_root_edges}
\title{Smooth the edges of a root binary image}
\usage{
smooth_root_edges(img, kernel_shape = "disk", kernel_size = 3, iterations = 1)
}
\arguments{
\item{img}{A binary image of root systems (imager cimg object)}

\item{kernel_shape}{Shape of the kernel: "square", "diamond", or "disk"}

\item{kernel_size}{Size of the kernel for morphological operations (odd integer). Use larger kernels for higher image resolution}

\item{iterations}{Number of iterations for the smoothing process}
}
\value{
A smoothed binary image
}
\description{
Smooth the edges of a root binary image
}
\examples{

data("seg_Oulanka2023_Session01_T067")
img <- seg_Oulanka2023_Session01_T067
# Try different kernel shapes
smoothed_square <- smooth_root_edges(img, kernel_shape = "square", kernel_size = 3)
smoothed_diamond <- smooth_root_edges(img, kernel_shape = "diamond", kernel_size = 3)
smoothed_disk <- smooth_root_edges(img, kernel_shape = "disk", kernel_size = 3)
plot(smoothed_disk)

}
\keyword{internal}
