% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zonestats3.R
\name{tube_coloration}
\alias{tube_coloration}
\title{Calculate Image Coloration Metrics}
\usage{
tube_coloration(img, r = 0.2126, g = 0.7152, b = 0.0722)
}
\arguments{
\item{img}{Three-band raster (RGB) or path to image.}

\item{r}{Red channel luminosity weight. Default follows ITU-R BT.709.}

\item{g}{Green channel luminosity weight.}

\item{b}{Blue channel luminosity weight.}
}
\value{
A data frame with columns: rcc, gcc, bcc, hue, saturation,
  luminosity, red, green, blue.
}
\description{
Calculate Image Coloration Metrics
}
\examples{
data(rgb_Oulanka2023_Session03_T067)
img <- terra::rast(rgb_Oulanka2023_Session03_T067)
colorvector <- tube_coloration(img)
}
