% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TurnoverDPC_data}
\alias{TurnoverDPC_data}
\title{Root Turnover Analysis Data}
\format{
A RasterBrick object with dimensions:
  \itemize{
    \item 2550 columns (width)
    \item 2273 rows (height)
    \item 3 layers (decay, growth, persistent)
  }
}
\source{
Original minirhizotron images acquired by Johannes Cunow
and Gesche Blume-Werry at Oulanka Research Station, Finland (June 2023).
Images were processed using RootDetector.
}
\usage{
data(TurnoverDPC_data)
}
\description{
Root turnover analysis from a sedge fen in northern Finland, comparing
root presence between two time points (June 2023 and October 2023) using
RootDetector root tracking feature.
}
\details{
Multi-layer representation of root dynamics where:
\itemize{
  \item Layer 1: Root decay (value = 1)
    \itemize{
      \item Roots that disappeared between time points
      \item Foreign objects (e.g., tape)
      \item Persistent roots
    }
  \item Layer 2: Root growth (value = 1)
    \itemize{
      \item New roots that appeared
      \item Persistent roots
    }
  \item Layer 3: Persistent roots (value = 1)
    \itemize{
      \item Roots present in both time points
    }
}
Background is represented as 0 in all layers.
}
\examples{
\dontrun{
  data(TurnoverDPC_data)
  img = terra::rast(TurnoverDPC_data)
  # Plot individual layers
  terra::plot(img, maxcell = Inf)
  # Calculate turnover statistics
  decay <- sum(terra::values(img[[1]]) == 1)
  growth <- sum(terra::values(img[[2]]) == 1)
  persistent <- sum(terra::values(img[[3]]) == 1)
}
}
\references{
Peters B, Blume-Werry G, Gillert A, et al. (2023)
RootDetector: a convolutional neural network for root detection.
Scientific Reports 13:1399. https://doi.org/10.1038/s41598-023-28400-x
}
\author{
Johannes Cunow
}
\keyword{datasets}
